Found a workaround for StackMapTable exception
The JVM verifies classes using a StackMapTable. Unfortunately, BCEL does
not update the offsets in this table when the code for a method is
changed, so the JVM is unhappy when certain code constructs are
instrumented (e.g. if statements or try/catch blocks). It is possible to
manually update the StackMapTable offsets, but it's not easy. For now,
the workaround is to simply use "-XX:-UseSplitVerifier" as a JVM option.