Major overhaul of the framework
Converted the project into an Eclipse plugin and added a builder and
launcher. The builder modifies class files and persists them to disk,
instead of the old way of dynamic modifying class files. The launcher
simply launches ThreadHarmony.class on the test class.

Implemented a way to dynamically generate all possible interleavings of
threads.

Added functionality to acquire shared variable information. Interleaves
are now only before instructions that access shared variables.
Currently, a shared variable is defined as a class field, but this may
change later on.